Layerize files
