+++ apps/core/settings.py
@@ -66,8 +66,8 @@ INSTALLED_APPS = [
     "data.streams",
     "data.trips",
     "data.post",
+    "data.files",
     "indieweb",
-    "files",
     "wordpress",
     "plugins",
     "interfaces",
