+++ apps/core/urls.py
@@ -6,11 +6,10 @@ from . import views
 
 urlpatterns = [
     path("a/", include("interfaces.dashboard.urls")),
-    path("a/", include("files.admin_urls")),
     path("a/", include("plugins.urls")),
     path("a/wordpress/", include("wordpress.urls", namespace="wordpress")),
     path("a/", include("indieweb.urls", namespace="indieweb")),
-    path("files/", include("files.urls")),
+    path("files/", include("interfaces.public.files.urls")),
     path("webmention/", include("webmention.urls")),
     path("admin/", admin.site.urls),
     path("auth/", include("django.contrib.auth.urls")),
