--- apps/files/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class FilesConfig(AppConfig):
-    name = "files"
