+++ apps/indieweb/application/webmentions.py
@@ -8,8 +8,8 @@ from data.post import models as post_models
 from data.post.models import TPost
 from django.conf import settings
 from django.utils.timezone import now
+from domain.files.utils import extract_uuid_from_url
 from domain.post import queries as post_queries
-from files.utils import extract_uuid_from_url
 from indieweb import models as indieweb_models
 from indieweb.constants import MPostKinds
 from webmention import models as webmention_models
