+++ apps/indieweb/utils.py
@@ -9,10 +9,10 @@ from typing import List, Optional, Union
 import requests
 from bs4 import BeautifulSoup
 from bs4.element import Tag
+from data.files.models import TFile
 from django.template.loader import render_to_string
-from files.forms import MediaUploadForm
-from files.images import bytes_as_upload_image
-from files.models import TFile
+from domain.images.images import bytes_as_upload_image
+from interfaces.public.files.forms import MediaUploadForm
 from mf2util import (
     _find_all_entries,
     classify_comment,
