+++ apps/indieweb/views.py
@@ -8,7 +8,6 @@ from django.shortcuts import get_object_or_404, redirect, render
 from django.utils import timezone
 from django.views.decorators.csrf import csrf_exempt
 from django.views.decorators.http import require_GET, require_POST
-from files.forms import MediaUploadForm
 from indieweb.application import micropub as micropub_app
 from indieweb.application import webmentions as webmention_app
 from indieweb.application.location import location_to_pointfield_input
@@ -22,6 +21,7 @@ from interfaces.dashboard.entry.forms import (
     TLocationModelForm,
     TSyndicationModelForm,
 )
+from interfaces.public.files.forms import MediaUploadForm
 from rest_framework import status
 from rest_framework.decorators import api_view
 from rest_framework.response import Response
