+++ apps/interfaces/common/forms/fields/_trix.py
@@ -1,7 +1,7 @@
 from bs4 import BeautifulSoup
+from data.files.constants import PICTURE_FORMATS
 from django import forms
 from django.template.loader import render_to_string
-from files.constants import PICTURE_FORMATS
 
 from ..widgets import TrixEditor
