+++ apps/interfaces/dashboard/urls.py
@@ -5,6 +5,7 @@ from plugins import pool
 urlpatterns = [
     path("", include("interfaces.dashboard.entry.urls")),
     path("", include("interfaces.dashboard.trips.urls")),
+    path("", include("interfaces.dashboard.files.urls")),
     path("", include("interfaces.dashboard.top.urls", namespace="post")),
 ]
