+++ apps/wordpress/views.py
@@ -13,8 +13,7 @@ from django.template.defaultfilters import linebreaks_filter, safe
 from django.urls import reverse
 from django.utils.decorators import method_decorator
 from django.views.generic import CreateView, ListView
-from files.forms import MediaUploadForm
-from files.images import bytes_as_upload_image
+from domain.images.images import bytes_as_upload_image
 from indieweb.application.location import location_to_pointfield_input
 from indieweb.utils import download_image, render_attachment
 from interfaces.dashboard.entry.forms import (
@@ -32,6 +31,7 @@ from interfaces.dashboard.entry.forms import (
     UpdateReplyForm,
     UpdateStatusForm,
 )
+from interfaces.public.files.forms import MediaUploadForm
 from turbo_response import TurboFrame, redirect_303
 
 from . import extract
